+++ Documents/Cpp-code/random_direction.cpp
@@ -7,11 +7,11 @@
 using namespace std;
 
 //Tested
-array<double, 3> randomDirection()
+array<long double, 3> randomDirection()
 {
-	double u = randomUniformDoubleClosed()*2.0 -1.0;
-	double theta = randomUniformDoubleOpen()*2.0*pi;
-	array<double, 3> vec = {sqrt(1.0 - pow(u,2.0))*cos(theta), sqrt(1.0 - pow(u,2.0))*sin(theta), u};
+	long double u = randomUniformDoubleClosed()*2.0 -1.0;
+	long double theta = randomUniformDoubleOpen()*2.0*pi;
+	array<long double, 3> vec = {sqrt(1.0 - pow(u,2.0))*cos(theta), sqrt(1.0 - pow(u,2.0))*sin(theta), u};
 	// Normalise
 	vec = normalise(vec);
 	return vec;
