+++ Documents/Cpp-code/random_direction.h
@@ -2,6 +2,6 @@
 #define RANDOM_DIRECTION_H
 
 //Tested
-std::array<double,3>  randomDirection();
+std::array<long double,3>  randomDirection();
 
 #endif
