+++ Documents/Cpp-code/random_numbers.cpp
@@ -3,27 +3,27 @@
 #include <limits>
 using namespace std;
 
-const double DBL_MAX = numeric_limits<double>::max();
+const long double DBL_MAX = numeric_limits<long double>::max();
 
 random_device rd;
 mt19937 gen(rd());
-uniform_real_distribution<double> closed_distribution(0.0, nextafter(1.0, DBL_MAX));
-uniform_real_distribution<double> open_distribution(0.0, 1.0);
+uniform_real_distribution<long double> closed_distribution(0.0, nextafter(1.0, DBL_MAX));
+uniform_real_distribution<long double> open_distribution(0.0, 1.0);
 
 
-double randomUniformDoubleClosed()
+long double randomUniformDoubleClosed()
 {
 	return closed_distribution(gen);
 }
 
 
-double randomUniformDoubleOpen()
+long double randomUniformDoubleOpen()
 {
 	return open_distribution(gen);
 }
 
-double randomExponential(double rate)
+long double randomExponential(long double rate)
 {
-	exponential_distribution<double> exp_distribution(rate);
+	exponential_distribution<long double> exp_distribution(rate);
 	return exp_distribution(gen);
 }
