+++ Documents/Cpp-code/random_numbers.h
@@ -2,12 +2,12 @@
 #define RANDOM_NUMBERS_H
 
 //Indirectly tested through random_direction
-double randomUniformDoubleClosed();
+long double randomUniformDoubleClosed();
 
 //Indirectly tested through random_direction
-double randomUniformDoubleOpen();
+long double randomUniformDoubleOpen();
 
 //Tested
-double randomExponential(double rate=1.0);
+long double randomExponential(long double rate=1.0);
 
 #endif
